requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-cloud-vision==3.4.5
python-multipart==0.0.6
pillow==10.1.0
pdf2image==1.16.3
python-dotenv==1.0.0
aiofiles==23.2.1
requests==2.31.0 